Fixed Energy Usage
I have changed the Cockpit class so that it will require the user to use
a initialization method to create Cockpit objects. This will allow for
things like screens to be added more easily. However, it will need to be
implemented for each Cockpit subclass, so a better method may be
required, such as using Decorator for Cockpit objects and not requiring
subclasses for the Cockpit class.